Add Fit class __getitem__
Allows users to get parameters via `__getitem__`. For example,
if a user has a Fit instance `fit` which has a matrix parameter `beta`
which has shape (3, 2) in Stan, `fit["beta"]` returns an array
with shape (3, 2, num_samples * num_chains).

This commit changes the shape of the array which is used to store
draws internally.