feat: Provide user feedback during compilation
Compiling the Stan program takes time. Give the user some indication of
what is happening.

Note that we cannot use a progress indicator here because httpstan
redirects everything sent to stderr to /dev/null to avoid overwhelming
the user with a flood of compiler info and warning messages. Progress
indicators can be used elsewhere.

Closes #111