fix: add a scripts/test script
Add a scripts/test script to make clearer how we test the
package. Previously the invocation of pytest was
only "documented" in our CI configuration. Now
it is "documented" in the repository proper in the form of a script.
With luck, anyone thinking to run `scripts/check` will
notice `scripts/test`.