feat: Add plugin system
Other developers can change the behavior of PyStan by writing a plugin.
Plugin developers should create a class which subclasses
`stan.plugins.PluginBase`. This class must be referenced in their
package's entry points section.

This feature is added with a particular use in mind. A developer now has
the ability run HMC diagnostics when sampling has completed and alert
the user if any problems are detected.

Closes #129