feat: Pin httpstan to major.minor version for reproducibility
Pin httpstan to major.minor version for reproducibility.
With this change, a given version of pystan will use a specific version
of Stan.  Previously a single version of pystan can use many different
versions of httpstan. Because the version of Stan is tied to the version
of httpstan, pystan could end up using a wide range of Stan versions.
Such behavior is inimical to reproducible research.

Closes #241