Output sampling info only when iterations change
Add some logic to keep track of previous number of iterations
across all chains and only print updates when that number changes.
The previous logic would print every time the status was polled.
This caused issues in Jupyter, which did not have the same benefit
of being able to clear the output as in a terminal, resulting in
the Jupyter interface being cluttered with thousands of sampling
updates on long-running models.