docs: Avoid suggesting support for numpy types
Remove the suggestion that the `Data` type accepts
`ndarray`s. (These are still unofficially supported.)

This simplifies the type, making it easier to read. This may help users
avoid assuming that arbitrary numpy types will be accepted as data.

PyStan barely depends on numpy. In the future, the dependency may be
removed. (NumPy is a serious dependency, wheel is ~16MB.) In the
meantime, users should be encouraged to interact with PyStan as if it
the package did not know about the existence of numpy.